Refactored reversi example into separate modules
