+++ examples/reversi/main.rs
@@ -4,39 +4,14 @@
 
 extern crate turtle;
 
+mod board;
+
 use std::f64::consts::PI;
 
-use turtle::{Turtle, Event, Color};
+use turtle::{Turtle, Event};
 use turtle::event::{MouseButton};
 
-/// None - empty tile
-/// Some(Piece::A) - occupied by piece A
-/// Some(Piece::B) - occupied by piece B
-///
-/// Each array in Board is a row of the board
-type Board = [[Option<Piece>; 8]; 8];
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum Piece {
-    A,
-    B,
-}
-
-impl Piece {
-    fn other(self) -> Self {
-        match self {
-            Piece::A => Piece::B,
-            Piece::B => Piece::A,
-        }
-    }
-
-    fn color(self) -> Color {
-        match self {
-            Piece::A => "#f44336".into(),
-            Piece::B => "#2196F3".into(),
-        }
-    }
-}
+use board::{Board, Piece};
 
 #[derive(Debug, Clone)]
 struct Dimensions {
